1) Backward compatibility changes 2) Tests
