+++ Source/EasyNetQ.Tests/Integration/Scheduling/PartyInvitation.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace EasyNetQ.Tests.Integration.Scheduling
+{
+    [Serializable]
+    public class PartyInvitation
+    {
+        public string Text { get; set; }
+        public DateTime Date { get; set; }
+    }
+}
