+++ Source/EasyNetQ/Scheduling/BusExtensions.cs
@@ -12,6 +12,7 @@ namespace EasyNetQ.Scheduling
 
         /// <summary>
         /// Schedule a message to be published at some time in the future.
+        /// This required the EasyNetQ.Scheduler service to be running.
         /// </summary>
         /// <typeparam name="T">The message type</typeparam>
         /// <param name="bus">The IBus instance to publish on</param>
@@ -19,20 +20,21 @@ namespace EasyNetQ.Scheduling
         /// <param name="message">The message to response with</param>
         public static void FuturePublish<T>(this IBus bus, DateTime futurePublishDate, T message) where T : class
         {
-            bus.Scheduler().FuturePublish(futurePublishDate, message);
+            bus.Scheduler().FuturePublishAsync(futurePublishDate, message).Wait();
         }
 
         /// <summary>
         /// Schedule a message to be published at some time in the future.
+        /// This required the EasyNetQ.Scheduler service to be running.
         /// </summary>
         /// <typeparam name="T">The message type</typeparam>
         /// <param name="bus">The IBus instance to publish on</param>
         /// <param name="futurePublishDate">The time at which the message should be sent (UTC)</param>
         /// <param name="cancellationKey">An identifier that can be used with CancelFuturePublish to cancel the sending of this message at a later time</param>
         /// <param name="message">The message to response with</param>
-        public static void FuturePublish<T>(this IBus bus, DateTime futurePublishDate, T message, string cancellationKey) where T : class
+        public static void FuturePublish<T>(this IBus bus, DateTime futurePublishDate, string cancellationKey, T message) where T : class
         {
-            bus.Scheduler().FuturePublish(futurePublishDate, message, cancellationKey);
+            bus.Scheduler().FuturePublishAsync(futurePublishDate, message, cancellationKey).Wait();
         }
 
         /// <summary>
@@ -44,7 +46,7 @@ namespace EasyNetQ.Scheduling
         /// <param name="message">The message to response with</param>
         public static void FuturePublish<T>(this IBus bus, TimeSpan messageDelay, T message) where T : class
         {
-            bus.Scheduler().FuturePublish(messageDelay, message);
+            bus.Scheduler().FuturePublishAsync(messageDelay, message).Wait();
         }
 
         /// <summary>
@@ -54,11 +56,12 @@ namespace EasyNetQ.Scheduling
         /// <param name="cancellationKey">The identifier that was used when originally scheduling the message with FuturePublish</param>
         public static void CancelFuturePublish(this IBus bus, string cancellationKey)
         {
-            bus.Scheduler().CancelFuturePublish(cancellationKey);
+            bus.Scheduler().CancelFuturePublishAsync(cancellationKey).Wait();
         }
 
         /// <summary>
         /// Schedule a message to be published at some time in the future.
+        /// This required the EasyNetQ.Scheduler service to be running.
         /// </summary>
         /// <typeparam name="T">The message type</typeparam>
         /// <param name="bus">The IBus instance to publish on</param>
@@ -71,13 +74,14 @@ namespace EasyNetQ.Scheduling
 
         /// <summary>
         /// Schedule a message to be published at some time in the future.
+        /// This required the EasyNetQ.Scheduler service to be running.
         /// </summary>
         /// <typeparam name="T">The message type</typeparam>
         /// <param name="bus">The IBus instance to publish on</param>
         /// <param name="futurePublishDate">The time at which the message should be sent (UTC)</param>
         /// <param name="cancellationKey">An identifier that can be used with CancelFuturePublish to cancel the sending of this message at a later time</param>
         /// <param name="message">The message to response with</param>
-        public static Task FuturePublishAsync<T>(this IBus bus, DateTime futurePublishDate, T message, string cancellationKey) where T : class
+        public static Task FuturePublishAsync<T>(this IBus bus, DateTime futurePublishDate, string cancellationKey, T message) where T : class
         {
             return bus.Scheduler().FuturePublishAsync(futurePublishDate, message, cancellationKey);
         }
